Finish migration to new GPFifo processor
